Set up just how I like it

+++ lua/plugins/mason.lua
@@ -3,7 +3,7 @@ return {
     "mason-org/mason.nvim",
     opts = {
       ui = {
-        border = "double",
+        border = "single",
       },
     },
   },
